+++ Changelog.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2022-01-18
+### Changed
+- Changed behaviour of `Ar::filter` and `Ar::uniqueValues` to automatically preserve keys only when logical based on the new `array_is_list` function. A polyfill for this function is loaded for lower PHP versions.
+### Removed
+- Removed `Ar::filterValues` use `Ar::filter`
+- Removed `Ar::uniqueValues` use `Ar::unique`
+- Removed `Ar::new` use `Ar::wrap`
+
 ## [0.14.0] - 2021-03-25
 - Use `iterable` typehint everywhere
 - Add `Ar::merge`.